pyproject.toml
[project]
name = "redes-lab8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kafka-python>=2.2.15",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
]
